remoção da lógico no componente, adicionar no arquivo .js apropriado e importar no componente.
